✨ Impl getAllHouseDeals Func

+++ SpringBoot_happyhouse/src/main/java/com/ssafy/web/vo/HouseDealVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class HouseDealVO {
-	private int no;
+	private BigInteger no;
 	
 	private String dealAmount;
 	private int dealYear;
